Latest OpenSearch Metrics with OSD
