+++ src/test/java/org/opensearchmetrics/model/label/LabelDataTest.java
@@ -1,4 +1,4 @@
-package org.opensearchmetrics.model;
+package org.opensearchmetrics.model.label;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -15,7 +15,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.Mockito.when;
 
-public class MetricsDataTest {
+public class LabelDataTest {
 
     @Mock
     ObjectMapper objectMapper;
@@ -28,24 +28,26 @@ public class MetricsDataTest {
     @Test
     void toJson() throws JsonProcessingException {
         // Arrange
-        MetricsData metricsData = new MetricsData();
-        metricsData.setId("1");
-        metricsData.setCurrentDate("2024-03-15");
-        metricsData.setRepository("test-repo");
-        metricsData.setMetricName("bugs");
-        metricsData.setMetricCount(10L);
+        LabelData labelData = new LabelData();
+        labelData.setId("1");
+        labelData.setCurrentDate("2024-03-15");
+        labelData.setRepository("test-repo");
+        labelData.setLabelName("bug");
+        labelData.setLabelIssueCount(10L);
+        labelData.setLabelPullCount(5L);
 
         Map<String, Object> expectedData = new HashMap<>();
         expectedData.put("id", "1");
         expectedData.put("current_date", "2024-03-15");
         expectedData.put("repository", "test-repo");
-        expectedData.put("metric_name", "bugs");
-        expectedData.put("metric_count", 10L);
+        expectedData.put("label_name", "bug");
+        expectedData.put("label_issue_count", 10L);
+        expectedData.put("label_pull_count", 5L);
 
         when(objectMapper.writeValueAsString(expectedData)).thenReturn("expectedJson");
 
         // Act
-        String actualJson = metricsData.toJson(objectMapper);
+        String actualJson = labelData.toJson(objectMapper);
 
         // Assert
         assertEquals("expectedJson", actualJson);
@@ -54,17 +56,18 @@ public class MetricsDataTest {
     @Test
     void getJson() throws JsonProcessingException {
         // Arrange
-        MetricsData metricsData = new MetricsData();
-        metricsData.setId("1");
-        metricsData.setCurrentDate("2024-03-15");
-        metricsData.setRepository("test-repo");
-        metricsData.setMetricName("bugs");
-        metricsData.setMetricCount(10L);
+        LabelData labelData = new LabelData();
+        labelData.setId("1");
+        labelData.setCurrentDate("2024-03-15");
+        labelData.setRepository("test-repo");
+        labelData.setLabelName("bug");
+        labelData.setLabelIssueCount(10L);
+        labelData.setLabelPullCount(5L);
 
         when(objectMapper.writeValueAsString(anyMap())).thenReturn("expectedJson");
 
         // Act
-        String actualJson = metricsData.getJson(metricsData, objectMapper);
+        String actualJson = labelData.getJson(labelData, objectMapper);
 
         // Assert
         assertEquals("expectedJson", actualJson);
@@ -73,16 +76,17 @@ public class MetricsDataTest {
     @Test
     void getJson_WithJsonProcessingException() throws JsonProcessingException {
         // Arrange
-        MetricsData metricsData = new MetricsData();
-        metricsData.setId("1");
-        metricsData.setCurrentDate("2024-03-15");
-        metricsData.setRepository("test-repo");
-        metricsData.setMetricName("bugs");
-        metricsData.setMetricCount(10L);
+        LabelData labelData = new LabelData();
+        labelData.setId("1");
+        labelData.setCurrentDate("2024-03-15");
+        labelData.setRepository("test-repo");
+        labelData.setLabelName("bug");
+        labelData.setLabelIssueCount(10L);
+        labelData.setLabelPullCount(5L);
 
         when(objectMapper.writeValueAsString(anyMap())).thenThrow(JsonProcessingException.class);
 
         // Act and Assert
-        assertThrows(RuntimeException.class, () -> metricsData.getJson(metricsData, objectMapper));
+        assertThrows(RuntimeException.class, () -> labelData.getJson(labelData, objectMapper));
     }
 }
